// angular-frontend/src/app/interfaces/vehicledemanded.ts
export class VehicleDemanded {
  id: number;
  vehicleNumber: string;
  driverName: string;
  vehicleType: string;
  capacity: number;
  costPerSlot: number;
  vehicleStatus: string;
  // slot: string,
  // date: string;
  slot1: string;
  slot2: string;
  slot3: string;
  companyName: string;
  slot: string;
  retailerId: number;
}